Refactor: Post/Comment 도메인 CQRS 패턴 분리
