+++ src/main/java/com/example/backend/domain/comment/controller/CommentQueryController.java
@@ -1,0 +1,32 @@
+package com.example.backend.domain.comment.controller;
+
+import com.example.backend.domain.comment.dto.CommentResponse;
+import com.example.backend.domain.comment.service.CommentQueryService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/posts/{postId}/comments")
+public class CommentQueryController {
+    private final CommentQueryService commentQueryService;
+
+    @Operation(summary = "댓글 조회", description = "특정 게시글의 댓글을 조회합니다.")
+    @ApiResponse(responseCode = "200", description = "댓글 조회 성공")
+    @GetMapping
+    public ResponseEntity<Page<CommentResponse>> getComments(
+            @PathVariable Long postId,
+            @PageableDefault(size = 10, sort = "createdAt") Pageable pageable){
+        Page<CommentResponse> comments = commentQueryService.getComments(postId, pageable);
+        return ResponseEntity.ok(comments);
+    }
+}
