+++ src/main/java/com/example/backend/domain/post/controller/PostCommandController.java
@@ -3,16 +3,13 @@ package com.example.backend.domain.post.controller;
 import com.example.backend.domain.post.dto.PostCreateRequest;
 import com.example.backend.domain.post.dto.PostResponse;
 import com.example.backend.domain.post.dto.PostUpdateRequest;
-
-import com.example.backend.domain.post.service.PostService;
+import com.example.backend.domain.post.service.command.PostCommandService;
+import com.example.backend.domain.post.service.command.PostDeleteService;
+import com.example.backend.domain.post.service.command.PostUpdateService;
 import com.example.backend.domain.security.adapter.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -24,9 +21,10 @@ import java.util.List;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/posts")
-public class PostController {
-
-    private final PostService postService;
+public class PostCommandController {
+    private final PostCommandService postCommandService;
+    private final PostUpdateService postUpdateService;
+    private final PostDeleteService postDeleteService;
 
     @Operation(summary = "이미지 포함 게시글 생성", description = "이미지를 포함한 게시글을 작성합니다.")
     @ApiResponse(responseCode = "200", description = "게시글 생성 성공")
@@ -37,25 +35,8 @@ public class PostController {
             @RequestPart("images") List<MultipartFile> images
     ) {
         Long userId = userDetails.getUserId();
-        PostResponse response = postService.createPost(userId, postCreateRequest, images);
+        PostResponse response = postCommandService.createPost(userId, postCreateRequest, images);
         return ResponseEntity.ok(response);
-    }
-
-    @Operation(summary = "게시글 조회", description = "특정 가게별 게시글을 조회합니다.")
-    @ApiResponse(responseCode = "200", description = "게시글 조회 성공")
-    @GetMapping("/{storeId}")
-    public ResponseEntity<Page<PostResponse>> getPostsByStore(
-            @PathVariable Long storeId,
-            @PageableDefault(size = 10, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
-        return ResponseEntity.ok(postService.getPostsByStore(storeId, pageable));
-    }
-
-    @Operation(summary = "게시글 세부 조회", description = "특정 게시글을 조회합니다.")
-    @ApiResponse(responseCode = "200", description = "게시글 조회 성공")
-    @GetMapping("/{storeId}/{postId}")
-    public ResponseEntity<PostResponse> getDetailPosts(@PathVariable Long storeId,
-                                                       @PathVariable Long postId) {
-        return ResponseEntity.ok(postService.getPostDetail(postId));
     }
 
     @Operation(summary = "게시글 수정", description = "특정 게시글을 수정합니다.")
@@ -69,7 +50,7 @@ public class PostController {
             @RequestPart("postUpdateRequest") PostUpdateRequest postUpdateRequest,
             @RequestPart(value = "images", required = false) List<MultipartFile> images
     ) {
-        PostResponse updated = postService.updatePost(postId, userDetails.getUserId(), postUpdateRequest, images);
+        PostResponse updated = postUpdateService.updatePost(postId, userDetails.getUserId(), postUpdateRequest, images);
         return ResponseEntity.ok(updated);
     }
 
@@ -80,7 +61,9 @@ public class PostController {
             @PathVariable Long postId,
             @PathVariable Long storeId,
             @AuthenticationPrincipal CustomUserDetails userDetails) {
-        postService.deletePost(postId, userDetails.getUserId());
+        postDeleteService.deletePost(postId, userDetails.getUserId());
         return ResponseEntity.noContent().build();
     }
+
+
 }
